Add files via upload
These two classes work hand-in-hand. 

nasdaq_2016 takes constituents of the Nasdaq from 2016 with the constituents each year up until the present.  It is used to make sure that there is no survivorship Bias. 

Casacade_Momentum is a strategy that simply takes the past 12 months Nasdaq returns top 50, of the top 50, the past 6 months return and takes the top 30, and of the top 30 takes the three month return and takes the top 10 stocks.